improve taken system and window management
The taken system was removed and put into the duty of the frames. Frames
can now be stashed and restored (popped from the stash) at a later time.

Now the `focus_frame` is allowed to be a parent frame, meaning actions
can be done on a group of frames. This was only really added because
when a monitor turns off, the frames and windows have to be stored
somewhere and this makes it easy to restore them after the monitor is
turned back on.

Windows are now managed differently. The `WM_STATE` client message is
now handled so windows can get set into fullscreen (or out of
fullscreen). The `WM_STATE` property is managed for fullscreen windows
now. Also no longer can a window change its window mode other than
through a `WM_STATE` client message. The user is in full control.

More client messages where added, see
`src/event.c:handle_client_message`.

Logging was drastically improved and made consistent across the board.
The `LOG` function can still be used like before but it supports
additional printf-format specifiers now, see
`include/log.h:formatted_log`.

Discord where the window is bugged after being closed has a bug that
will be fixed in the next commit.