improve floating, dock sizing and add to configuration
Moving from one frame to another while crossing a monitor edge now
checks if there is a floating/fullscreen window covering the monitor
with a specific percentage (configurable through `general.move_overlap`)
and then focuses the window (while still focusing the frame it's
supposed to).

Now there is the `[startup]` section which can be filled with actions to
run when the window manager starts. The configuration parser code was
cleaned up a lot as well.

Floating windows are now sized more sensibly as well as dock windows.