implement #19: Improve frame indicator message
Now a message showing where the current frame is only shown when there
is no window in that frame or there are no window borders.
Reasoning is simple: If there is already a mean (changing border colors)
to indicate the frame focus, then there should not be a second one.