fix #28: Old focus is set to pointer
The old focus frame/window is set to a pointer in `next_cycle()`. It
could in theory happen that this pointer is freed which is no issue
since it is used only for comparison. But what if the pointer is freed
and then an allocated pointer has the exact same address? To stop this
issue, the old focus frame/window is handled externally now when a
frame/window is destroyed.