implement #31: Toggling modes is incomplete
When toggling a mode to tiling and from tiling, the new configuration
options were not considered. Now an outer functions is both used in
`show_window()` and `set_window_mode()` because that code section was
exactly the same.

There are a few fixes in `src/action.c` as well. One is pretty
unrelated where exchanging frames hid the windows when it was not
supposed to.