clean up and fix bugs
This has no specific associated task, it is just general cleanup.

`utility.h` was cleaned up and an `APPROXIMATE_DIGITS()` macro was
added which is used twice.

The frames were not sized before running the startup actions, this is
now fixed. There were some other situations where the frames were not
sized properly, all fixed.

The management for _NET_WORKAREA was abandoned since it makes no sense
to manage it for multiple monitors.

The stash was fixed, the windows were not validated in
`unlink_frame_from_stash()`.

The `equalize_frame()` was again improved and now it should behave most
intuitively.