implement #36: add way to inject commands
Now commands can be injected by mapping an InputOnly window that has the
FENSTERCHEF_COMMAND property. This mapped window has to wait until this
property is deleted from the window before it is allowed to destroy
itself.

FENSTERCHEF_COMMAND is a text property (format 8) that can hold actions like in
a regular fensterchef configuration file.

Details are in fensterchef.c:run_external_command.

Fensterchef can be used with `-e`/`--command`, for example `fensterchef
-e 'run $TERMINAL'` to run a terminal. This does all described above
(mapping a window and waiting).