implement #41: Add actions to assignment
Now after an assignment can optionally be added a list of actions to
execute when a window matches the assignment.

The `SHOW-WINDOW` actions is added which is the same as `FOCUS-WINDOW`
but it only shows it where as the latter shows and focuses the window.

"generate/fensterchef.5.sh" was added, I forgot in a previous commit.