fix #52: Let more frames tank space
Now more frames are pushed away to gain space. The `get_XXX_frame()`
functions were adjusted to me more layout independent.

Few bugs were fixed.

A rectangle is now defined to be within a monitor most when its center
is within the monitor.
This will not make a difference in most situation but in the rare
instance when a window is centered within a smaller monitor but it
extends out more to a larger monitor.