fix #67: GVim not rendering properly
If GVim is spawned, it does not render properly, it is just black.
Resizing the window however fixes the issue.

The solution is to send a fake configure notify event to the window so
that it stays in a valid state.

This commit includes the addition of WM_STATE. I thought this would fix
the bug but it did not. I believe however this fixed bug #7 where
the discord application window hides itself again.