implement #75: Stable floating window positioning
Floating windows have a stable and predictable positioning now. When a
floating window is hidden now, it will always appear at the same (x,y)
position and will not be clipped or put on another monitor.

This reverts a past change not directly associated to a task.

The user can make a bind if it often happens that windows get "lost".
But this should not happen in a usual setup.