implement #72: Re-design configuration parser
This is a complete redesign of the configuration parser.

Now it is a very simple file format that is easy to highlight and
parse.  No features were omitted and it is as powerful as it was
originally.
The sections were completely removed, they are very obsolete.

The generation code was abandonded and will partially be rewritten for
the documentation and the vim syntax highlighting file that is not
included yet.