implement #87/#88: Improve bindings
Now we use a binding map for efficient key/button lookup.
The bindings are actions themself, simplifying a lot and allowing to set
bindings within bindings.  This however needs to fully implemented and
will be with task #85.

There is a lot of restructuring.  The configuration was moved to the
root again and the configuration directory was renamed to "parse".

There is quite a bit of rubble around.  The client list synchronizing
needs to be heavily improved.